R1: schema-literal type inference - the type system now reaches the columns
defineComponent/defineResource infer the value type from the schema
literal (const type param; ValueOf mapped type) - the hand-declared <S>
and its silent-drift hazard are gone. Component carries the schema as a
second defaulted param so downstream value-typed signatures did not
ripple. col() returns ColumnsOf<Sch> (concrete typed-array types per
field; field typos are compile errors). spawn is typed per-entry via a
schema-tuple signature (WriteOf makes defaulted fields optional,
required fields required; wrong/missing/bad-enum values rejected at the
call site - no NoInfer needed). readField is name-keyed with the return
type from the schema; public readEid deleted (subsumed; the internal
store validator remains). enumOf is generic over labels in both
overloads; field() types its default.

Truthful nullability: "string" reads as string | null, "key" as
EntityKey | null (null is a legal stored cell - encodeField/decodeField
round-trip it; confirmed by runtime probe). This flushed out a latent
example bug: contentLayer note-text guard checked !== undefined where
null was possible.

The example is the proof: every hand-declared component type and all 42
"as Float32Array/..." casts deleted; inference flows from the schema
literal to the hot loop. dist d.ts verified to carry the inference
through rollup-plugin-dts. +3 compile-time probe tests (259 total).

Per docs/review-part1.md R1 - the shape design.md always promised.

+++ examples/canvas-editor/src/app/camera.ts
@@ -59,10 +59,10 @@ export function zoomToFit(): void {
   let maxX = -Infinity;
   let maxY = -Infinity;
   worldRef.current.query(renderable).each((b) => {
-    const px = b.col(Position).x as Float32Array;
-    const py = b.col(Position).y as Float32Array;
-    const sw = b.col(Size).w as Float32Array;
-    const sh = b.col(Size).h as Float32Array;
+    const px = b.col(Position).x;
+    const py = b.col(Position).y;
+    const sw = b.col(Size).w;
+    const sh = b.col(Size).h;
     for (let i = 0; i < b.count; i++) {
       const r = b.rows[i];
       minX = Math.min(minX, px[r] - sw[r] / 2);
